pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jeffrey Nicholas <jefnic23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
fastapi-mail = "^1.4.1"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
aioboto3 = "^12.4.0"
email-validator = "^2.1.1"
python-multipart = "^0.0.9"
aiohttp = {extras = ["speedups"], version = "^3.9.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
